fix some review stuff
