+++ server/controllers/commonController.js
@@ -1,72 +1,17 @@
 import * as commonService from "../services/commonService.js";
 import * as voteService from "../services/voteService.js";
-import mongoose from "mongoose";
+import {
+    hasInvalidObjectId,
+    parseBoolean,
+    parseCsv,
+    parsePositiveInt,
+    parseSearchTypes as parseSearchTypesFromQuery,
+} from "../utils/inputParsers.js";
 
 const SEARCH_ALLOWED_TYPES = new Set(["discussion", "course"]);
 
-function parseCsv(value) {
-    if (!value) return [];
-
-    if (Array.isArray(value)) {
-        return value
-            .map((item) => String(item).trim())
-            .map((item) => item.replace(/^['"]+|['"]+$/g, ""))
-            .filter(Boolean);
-    }
-
-    if (typeof value !== "string") return [];
-
-    const raw = value.trim();
-    if (!raw) return [];
-
-    if (raw.startsWith("[") && raw.endsWith("]")) {
-        try {
-            const parsed = JSON.parse(raw);
-            if (Array.isArray(parsed)) {
-                return parsed
-                    .map((item) => String(item).trim())
-                    .map((item) => item.replace(/^['"]+|['"]+$/g, ""))
-                    .filter(Boolean);
-            }
-        } catch {
-            // fall through to CSV parsing
-        }
-    }
-
-    return raw
-        .split(/[\n,]/)
-        .map((item) => item.trim())
-        .map((item) => item.replace(/^['"]+|['"]+$/g, ""))
-        .filter(Boolean);
-}
-
-function hasInvalidObjectId(ids = []) {
-    return ids.some((id) => !mongoose.Types.ObjectId.isValid(id));
-}
-
 function parseSearchTypes(value) {
-    const parsed = parseCsv(value).map((item) => item.toLowerCase());
-    if (parsed.length === 0) return ["discussion", "course"];
-
-    const types = parsed.filter((type) => SEARCH_ALLOWED_TYPES.has(type));
-    if (types.length === 0) return null;
-    return types;
-}
-
-function parseBoolean(value) {
-    if (typeof value === "boolean") return value;
-    if (typeof value !== "string") return undefined;
-
-    const normalized = value.trim().toLowerCase();
-    if (["1", "true", "yes"].includes(normalized)) return true;
-    if (["0", "false", "no"].includes(normalized)) return false;
-    return undefined;
-}
-
-function parsePositiveInt(value, fallback) {
-    const parsed = Number.parseInt(value, 10);
-    if (!Number.isFinite(parsed) || parsed <= 0) return fallback;
-    return parsed;
+    return parseSearchTypesFromQuery(value, SEARCH_ALLOWED_TYPES, ["discussion", "course"]);
 }
 
 export async function search(req, res) {
