+++ server/utils/regex.js
@@ -1,0 +1,3 @@
+export function escapeRegex(value) {
+  return String(value).replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
+}
